Fix zsh incorrect zdotdir bug :rage4: (#1)
# Description
I was originally naive with my `.zshrc` config file implementation and I ended up modifying the `$ZDOTDIR` variable there, which caused child zsh processes to not source the `.zshrc` file when starting (e. g. the integrated terminal in VScode) since the `$ZDOTDIR` variable had the path as `~/.config/zsh` and there was no `.zshrc` file there. With these changes, we now have a `.zshenv` file in the home directory that is used to set env variables like `$ZDOTDIR`  while we have the `.zshrc` in our `$ZDOTDIR` directory :godmode:.

## What changed:

* Rename .zshrc to ..config/zsh/.zshrc

* Update and rename ..config/zsh/.zshrc to .config/zsh/zshrc

This file should be on the $ZDOTDIR folder, not in the home folder or else zsh child processes will not source it.

* Create .zshenv

This will set the $ZDOTDIR env variable correctly.

* Update README.md

* Rename zshrc to .zshrc

* Update README.md